waylandclipboard: Use a dedicated event queue
When using KSystemClipboard on Wayland, reading from QClipboard will
fail because the main thread is blocked waiting for data so the send
event is never dispatched. A Program can avoid this and always use
KSystemClipboard but this is pragmatically not always possible for
example in plasmashell.
Instead of using the default queue, use a dedicated one and disptach
it ourselves. This enables the later programs to use KSystemClipboard
in a different thread.
CCBUG:442521